# openapi/plystra.v0.0.1.yaml
components:
    schemas:
...
        ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission:
            properties:
                data:
                    $ref: '#/components/schemas/ApiOpenAPIPermission'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission:
            properties:
                data:
                    $ref: '#/components/schemas/ApiOpenAPIRolePermission'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIError:
            properties:
                audit_log_id:
                    example: audit_01hzyj6b6q8
                    type: string
                code:
                    example: VALIDATION_FAILED
                    type: string
                deny_code:
                    example: SCOPE_OUT_OF_BOUNDS
                    type: string
                details: {}
                message:
                    example: Request body is invalid JSON.
                    type: string
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
                trace_id:
                    example: trace_01hzyj6b6q8
                    type: string
            type: object
        ApiOpenAPIErrorEnvelope:
            properties:
                error:
                    $ref: '#/components/schemas/ApiOpenAPIError'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIPermission'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission:
            properties:
                data:
                    items:
                        $ref: '#/components/schemas/ApiOpenAPIRolePermission'
                    nullable: true
                    type: array
                pagination:
                    $ref: '#/components/schemas/ApiOpenAPIPagination'
                request_id:
                    example: req_01hzyj6b6q8
                    type: string
            type: object
...
        ApiOpenAPIPagination:
            properties:
                cursor:
                    nullable: true
                    type: string
                has_more:
                    example: false
                    type: boolean
                limit:
                    example: 50
                    type: integer
            type: object
        ApiOpenAPIPermission:
            properties:
                action:
                    type: string
                created_at:
                    format: date-time
                    type: string
                deleted_at:
                    format: date-time
                    nullable: true
                    type: string
                description:
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                resource:
                    type: string
                scope:
                    type: string
                status:
                    type: string
                updated_at:
                    format: date-time
                    type: string
            type: object
...
        ApiPermissionMutationRequest:
            properties:
                action:
                    type: string
                actor:
                    $ref: '#/components/schemas/AuthzActorContext'
                audit_space_id:
                    type: string
                description:
                    nullable: true
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                resource:
                    type: string
                scope:
                    type: string
                status:
                    nullable: true
                    type: string
            type: object
...
        ApiRolePermissionMutationRequest:
            properties:
                actor:
                    $ref: '#/components/schemas/AuthzActorContext'
                audit_space_id:
                    type: string
                id:
                    type: string
                metadata:
                    additionalProperties: {}
                    nullable: true
                    type: object
                permission_id:
                    type: string
                role_id:
                    type: string
            type: object
...
        AuthzActorContext:
            properties:
                binding_id:
                    type: string
                binding_status:
                    type: string
                member_display_name:
                    type: string
                member_id:
                    type: string
                member_status:
                    type: string
                relation_type:
                    type: string
                space_id:
                    type: string
                space_name:
                    type: string
                space_status:
                    type: string
                user_email:
                    type: string
                user_id:
                    type: string
                user_member_id:
                    type: string
                user_status:
                    type: string
            type: object
...
    securitySchemes:
        ApiKeyAuth:
            description: Scoped API key for server-to-server Core calls.
            in: header
            name: X-Plystra-API-Key
            type: apiKey
        BearerAuth:
            bearerFormat: opaque Plystra access token
            description: Use the access_token returned by POST /api/v1/auth/login or POST /api/v1/auth/refresh.
            scheme: bearer
            type: http
        MetricsTokenAuth:
            description: Dedicated metrics token when METRICS_TOKEN or PLYSTRA_METRICS_TOKEN is configured.
            in: header
            name: X-Plystra-Metrics-Token
            type: apiKey
info:
    description: Self-hosted Plystra Core API. Stable surfaces cover identity, authorization, resource registry, audit, admin control-plane, and scoped API keys; plugin, template, and Data Console routes are preview metadata surfaces.
    title: Plystra Core API
    version: 0.0.1
openapi: 3.0.3
paths:
...
    /api/v1/app-modules/{module_key}/permissions:
        get:
            operationId: listAppModulePermissions
            parameters:
                - in: path
                  name: module_key
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List app module permissions
            tags:
                - App Modules
...
    /api/v1/permissions:
        get:
            operationId: listPermissions
            parameters:
                - description: Maximum number of rows to return.
                  in: query
                  name: limit
                  schema:
                    description: Maximum number of rows to return.
                    maximum: 200
                    minimum: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List permissions
            tags:
                - Permissions
        post:
            operationId: createPermission
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiPermissionMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create a permission
            tags:
                - Permissions
    /api/v1/permissions/{permission_id}:
        get:
            operationId: getPermission
            parameters:
                - in: path
                  name: permission_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get a permission
            tags:
                - Permissions
        patch:
            operationId: updatePermission
            parameters:
                - in: path
                  name: permission_id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiPermissionMutationRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Update a permission
            tags:
                - Permissions
...
    /api/v1/plugins/{plugin_key}/permissions:
        get:
            operationId: listPluginPermissions
            parameters:
                - in: path
                  name: plugin_key
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIPermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List plugin permissions
            tags:
                - Plugins
...
    /api/v1/role-permissions:
        get:
            operationId: listRolePermissions
            parameters:
                - description: Maximum number of rows to return.
                  in: query
                  name: limit
                  schema:
                    description: Maximum number of rows to return.
                    maximum: 200
                    minimum: 1
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIListEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: List role-permission bindings
            tags:
                - Permissions
        post:
            operationId: createRolePermission
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiRolePermissionMutationRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission'
                    description: Created
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Create a role-permission binding
            tags:
                - Permissions
    /api/v1/role-permissions/{role_permission_id}:
        delete:
            operationId: deleteRolePermission
            parameters:
                - in: path
                  name: role_permission_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Revoke a role-permission binding
            tags:
                - Permissions
        get:
            operationId: getRolePermission
            parameters:
                - in: path
                  name: role_permission_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIEnvelopeGithubComPlystraCoreInternalApiOpenAPIRolePermission'
                    description: OK
                default:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiOpenAPIErrorEnvelope'
                    description: Error response.
            security:
                - BearerAuth: []
                - ApiKeyAuth: []
            summary: Get a role-permission binding
            tags:
                - Permissions
...
servers:
    - url: http://localhost:8080
...